standardize

+++ setup/email/email_jobs.sh
@@ -1,11 +1,23 @@
 #!/bin/bash
 
-#make it available for both test and prod in the future
+# cd /home/usr/personal_projects/jobs_tracker/
 
-#Figure out how to cd to relative path; need to delete this absolute strategy
-cd /home/suraj/personal_projects/jobs_tracker/
-source setenv.sh test
+if [[ $# -ne 2 ]]; then
+    echo "Two args required. First: run mode, second: absolute path for repo."
+    exit
+fi
 
-docker compose up database_test test_entrypoint -d
-docker exec test_box bash -c "python email_utils/send_email_utils.py"
-docker compose down
+cd $1
+source setenv.sh $2
+
+if [[ $2 == "test" ]]; then
+    docker compose up test_database test_entrypoint -d
+    docker exec test_box bash -c "python email_utils/send_email_utils.py"
+    docker compose down
+
+elif [[ $2 == "prod" ]]; then
+    docker compose up prod_database prod_entrypoint -d
+    docker exec test_box bash -c "python email_utils/send_email_utils.py"
+    docker compose down
+
+fi
